README: passing channel name with quotes
The reason why #channel doesn't work is because the running shell
see the string with the beginning # symbol and whatever comes after that
as a comment. One workaround is to quote the #channel name as a string
literal. Example: ./irc_client.py myusername '#perl'. This also applies
to myusername parameter if it contains # symbol.